# docs/swagger.yaml
basePath: /api/v1/
definitions:
  lib.Page:
    properties:
      first:
        description: indicate first value
        example: true
        type: boolean
      items:
        type: object
      last:
        description: indicate last value
        example: false
        type: boolean
      max_page:
        description: maximum pages for current schema
        example: 9
        type: integer
      page:
        description: current page, start from zero
        example: 0
        type: integer
      size:
        description: size per page, default `10`
        example: 10
        type: integer
      total:
        description: total data without filters
        example: 100
        type: integer
      total_pages:
        description: total pages
        example: 10
        type: integer
      visible:
        description: current length
        example: 10
        type: integer
    type: object
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
...
  model.Brand:
    properties:
      brand_code:
        example: XM
        type: string
      brand_name:
        example: Iphone
        type: string
    required:
    - brand_code
    - brand_name
    type: object
  model.BrandAPI:
    properties:
      brand_code:
        example: XM
        type: string
      brand_name:
        example: Iphone
        type: string
    required:
    - brand_code
    - brand_name
    type: object
...
  model.LoginAPI:
    properties:
      password:
        example: '@Password123'
        type: string
      remember:
        example: true
        type: boolean
      username:
        example: john.doe@mail.com
        type: string
    required:
    - password
    - username
    type: object
...
  model.RegistrationAPI:
    properties:
      confirm_password:
        example: '@Password123'
        type: string
      email:
        example: john.doe@mail.com
        type: string
      fullname:
        example: John doe
        type: string
      password:
        example: '@Password123'
        minLength: 8
        type: string
      referral_code:
        example: E6So5no5
        type: string
    required:
    - confirm_password
    - email
    - fullname
    - password
    type: object
  model.SendVerificationAccountAPI:
    properties:
      email:
        example: john.doe@mail.com
        type: string
    required:
    - email
    type: object
...
  model.UserAPI:
    properties:
      activated_at:
        format: date-time
        type: string
      asset_id:
        type: string
      email:
        type: string
      fullname:
        type: string
      is_activated:
        type: boolean
      is_owner:
        type: boolean
      is_password_system_generated:
        type: boolean
      last_login:
        format: date-time
        type: string
      mobile:
        type: string
      otp_enabled:
        type: boolean
      otp_expiration:
        format: date-time
        type: string
      password_expiration:
        format: date-time
        type: string
      password_last_change:
        format: date-time
        type: string
      reset_password_expiration:
        format: date-time
        type: string
      username:
        type: string
      verification_expiration:
        format: date-time
        type: string
    type: object
  model.VerificationAccountAPI:
    properties:
      email:
        example: john.doe@mail.com
        type: string
      verification_code:
        example: "7085"
        type: string
    required:
    - email
    - verification_code
    type: object
info:
  contact:
    email: erfanpratama321@gmail.com
    name: Irfan Pratama
  description: API Documentation
  termsOfService: https://dospecs.monstercode.net/en/guide/tnc.html
  title: User Services
  version: 1.0.0
paths:
...
  /account/login:
    post:
      description: Generate token
      parameters:
      - description: Payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.LoginAPI'
      responses:
        "200":
          description: Logedin
          schema:
            $ref: '#/definitions/model.LoginAPI'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - TokenKey: []
      summary: Generate token
      tags:
      - Account
  /account/register:
    post:
      description: Registration
      parameters:
      - description: Payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.RegistrationAPI'
      responses:
        "200":
          description: registered
          schema:
            $ref: '#/definitions/lib.Response'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - TokenKey: []
      summary: Registration
      tags:
      - Account
  /account/send-verify-account-code:
    post:
      description: Resend Code Verify account
      parameters:
      - description: Payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.SendVerificationAccountAPI'
      responses:
        "200":
          description: Account verified
          schema:
            $ref: '#/definitions/model.UserAPI'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - TokenKey: []
      summary: Resend Code Verify account
      tags:
      - Account
  /account/verify-account:
    post:
      description: Verify account
      parameters:
      - description: Payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.VerificationAccountAPI'
      responses:
        "200":
          description: Account verified
          schema:
            $ref: '#/definitions/model.UserAPI'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - TokenKey: []
      summary: Verify account
      tags:
      - Account
  /brands:
    get:
      consumes:
      - application/json
      description: List of Brand
      parameters:
      - description: Page number start from zero
        in: query
        name: page
        type: integer
      - description: Size per page, default `0`
        in: query
        name: size
        type: integer
      - description: Sort by field, adding dash (`-`) at the beginning means descending
          and vice versa
        in: query
        name: sort
        type: string
      - description: Select specific fields with comma separated
        in: query
        name: fields
        type: string
      - description: custom filters, see [more details](https://github.com/morkid/paginate#filter-format)
        in: query
        name: filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Brand
          schema:
            allOf:
            - $ref: '#/definitions/lib.Page'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/model.Brand'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: List of Brand
      tags:
      - Brand
    post:
      consumes:
      - application/json
      description: Create new Brand
      parameters:
      - description: Brand data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.BrandAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Brand data
          schema:
            $ref: '#/definitions/model.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Create new Brand
      tags:
      - Brand
  /brands/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Brand by id
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Brand by id
      tags:
      - Brand
    get:
      consumes:
      - application/json
      description: Get a Brand by id
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Brand data
          schema:
            $ref: '#/definitions/model.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Get a Brand by id
      tags:
      - Brand
    put:
      consumes:
      - application/json
      description: Update Brand by id
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Brand data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.BrandAPI'
      produces:
      - application/json
      responses:
        "200":
          description: Brand data
          schema:
            $ref: '#/definitions/model.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/lib.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Brand by id
      tags:
      - Brand
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  TokenKey:
    in: header
    name: x-Token
    type: apiKey
swagger: "2.0"